Move DraftailRichTextArea tests to draftail.test.js and decouple from telepath
