+++ client/src/entrypoints/admin/draftail.test.js
@@ -1,3 +1,7 @@
+import { createEditorStateFromRaw } from 'draftail';
+import { EditorState } from 'draft-js';
+import ReactTestUtils from 'react-dom/test-utils';
+
 require('./draftail');
 
 describe('draftail', () => {
@@ -163,5 +167,222 @@ describe('importing the module multiple times', () => {
     /* eslint-enable no-console */
     jest.clearAllMocks();
     jest.restoreAllMocks();
+  });
+});
+
+describe('DraftailRichTextArea', () => {
+  let boundWidget;
+  let inputElement;
+  let parentCapabilities;
+
+  const TEST_RAW = {
+    blocks: [
+      {
+        key: 't30wm',
+        type: 'unstyled',
+        depth: 0,
+        text: 'Test Bold Italic',
+        inlineStyleRanges: [
+          {
+            offset: 5,
+            length: 4,
+            style: 'BOLD',
+          },
+          {
+            offset: 10,
+            length: 6,
+            style: 'ITALIC',
+          },
+        ],
+        entityRanges: [],
+      },
+    ],
+    entityMap: {},
+  };
+  const TEST_VALUE = JSON.stringify(TEST_RAW);
+
+  beforeEach(() => {
+    // Create a placeholder to render the widget
+    document.body.innerHTML = '<div id="placeholder"></div>';
+
+    const widgetDef = new window.draftail.DraftailRichTextArea({
+      entityTypes: [
+        {
+          type: 'LINK',
+          icon: 'link',
+          description: 'Link',
+          attributes: ['url', 'id', 'parentId'],
+          allowlist: {
+            href: '^(http:|https:|undefined$)',
+          },
+        },
+        {
+          type: 'IMAGE',
+          icon: 'image',
+          description: 'Image',
+          attributes: ['id', 'src', 'alt', 'format'],
+          allowlist: {
+            id: true,
+          },
+        },
+      ],
+      enableHorizontalRule: true,
+      inlineStyles: [
+        {
+          type: 'BOLD',
+          icon: 'bold',
+          description: 'Bold',
+        },
+        {
+          type: 'ITALIC',
+          icon: 'italic',
+          description: 'Italic',
+        },
+      ],
+      blockTypes: [
+        {
+          label: 'H2',
+          type: 'header-two',
+          description: 'Heading 2',
+        },
+      ],
+    });
+    parentCapabilities = new Map();
+    parentCapabilities.set('split', { enabled: true, fn: jest.fn() });
+    parentCapabilities.set('addSibling', {
+      enabled: true,
+      getBlockMax: () => 5,
+      blockGroups: [
+        [
+          'Media',
+          [
+            {
+              name: 'image_block',
+              meta: {
+                icon: 'image',
+                label: 'Image',
+                blockDefId: 'blockdef-1234',
+                isPreviewable: true,
+                description: 'Full-width image',
+              },
+            },
+          ],
+        ],
+      ],
+      fn: jest.fn(),
+    });
+    const inputId = 'the-id';
+    boundWidget = widgetDef.render(
+      document.getElementById('placeholder'),
+      'the-name',
+      inputId,
+      TEST_VALUE,
+      parentCapabilities,
+      {
+        attributes: {
+          maxLength: 512,
+        },
+      },
+    );
+    inputElement = document.querySelector('#the-id');
+  });
+
+  test('it renders correctly', () => {
+    expect(document.querySelector('.Draftail-Editor__wrapper')).toBeTruthy();
+    expect(document.querySelector('input').value).toBe(TEST_VALUE);
+    expect(document.querySelector('input').maxLength).toBe(512);
+  });
+
+  test('getValue() returns the current value', () => {
+    expect(boundWidget.getValue()).toBe(TEST_VALUE);
+  });
+
+  test('getState() returns the current state', () => {
+    const state = createEditorStateFromRaw(TEST_RAW);
+    let retrievedState = boundWidget.getState();
+    // Ignore selection, which is altered from the original state by Draftail,
+    // (TODO: figure out why this happens)
+    // and decorator, which is added to by CommentableEditor
+    retrievedState = EditorState.acceptSelection(
+      retrievedState,
+      state.getSelection(),
+    );
+    retrievedState = EditorState.set(retrievedState, {
+      decorator: state.getDecorator(),
+    });
+    expect(retrievedState).toStrictEqual(state);
+  });
+
+  test('setState() changes the current state', () => {
+    const NEW_VALUE = {
+      blocks: [
+        {
+          key: 't30wm',
+          type: 'unstyled',
+          depth: 0,
+          text: 'New value',
+          inlineStyleRanges: [],
+          entityRanges: [],
+        },
+      ],
+      entityMap: {},
+    };
+    const NEW_STATE = createEditorStateFromRaw(NEW_VALUE);
+    boundWidget.setState(NEW_STATE);
+
+    let retrievedState = boundWidget.getState();
+    // Ignore selection, which is altered from the original state by Draftail,
+    // and decorator, which is added to by CommentableEditor
+    retrievedState = EditorState.acceptSelection(
+      retrievedState,
+      NEW_STATE.getSelection(),
+    );
+    retrievedState = EditorState.set(retrievedState, {
+      decorator: NEW_STATE.getDecorator(),
+    });
+    expect(retrievedState).toStrictEqual(NEW_STATE);
+  });
+
+  test('setInvalid() sets aria-invalid attribute', () => {
+    boundWidget.setInvalid(true);
+    expect(inputElement.getAttribute('aria-invalid')).toBe('true');
+    boundWidget.setInvalid(false);
+    expect(inputElement.getAttribute('aria-invalid')).toBeNull();
+  });
+
+  test('focus() focuses the text input', () => {
+    // focus happens on a timeout, so use a mock to make it happen instantly
+    jest.useFakeTimers();
+    boundWidget.focus();
+    jest.runAllTimers();
+    expect(document.activeElement).toBe(
+      document.querySelector('.public-DraftEditor-content'),
+    );
+  });
+
+  test('setCapabilityOptions for split updates the editor commands', () => {
+    ReactTestUtils.act(() =>
+      boundWidget.setCapabilityOptions('split', { enabled: false }),
+    );
+    expect(inputElement.draftailEditor.props.commands).toHaveLength(2);
+    ReactTestUtils.act(() =>
+      boundWidget.setCapabilityOptions('split', { enabled: true }),
+    );
+    expect(inputElement.draftailEditor.props.commands).toHaveLength(4);
+    expect(inputElement.draftailEditor.props.commands[3].items[0].type).toBe(
+      'split',
+    );
+  });
+
+  test('configures the block chooser based on siblings capability', () => {
+    expect(inputElement.draftailEditor.props.commands[2].items[0]).toEqual(
+      expect.objectContaining({
+        icon: 'image',
+        label: 'Image',
+        blockDefId: 'blockdef-1234',
+        isPreviewable: true,
+        description: 'Full-width image',
+      }),
+    );
   });
 });
